Move progress fetching code from Student model to request handler and fix
